work day 2 27/5/2025

+++ app/src/main/java/com/example/audioplayerdemo3/audioService/AudioServiceHandler.kt
@@ -123,7 +123,7 @@ sealed class PlayerEvent{
     object Forward: PlayerEvent()
     object SeekTo: PlayerEvent()
     object Stop: PlayerEvent()
-    data class UpdateProgress(val newProgress: Long): PlayerEvent()
+    data class UpdateProgress(val newProgress: Float): PlayerEvent()
 }
